add postgres eventstore and readrepository
in the process, refactor tests and examples so that
the postgres eventstore and readrepository could reuse existing
tests. then, did the same for all existing tests.
