+++ eventstore_tracing_test.go
@@ -1,96 +1,10 @@
-// Copyright (c) 2014 - Max Persson <max@looplab.se>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package eventhorizon
 
 import (
 	. "gopkg.in/check.v1"
 )
 
-var _ = Suite(&MemoryEventStoreSuite{})
 var _ = Suite(&TraceEventStoreSuite{})
-
-type MemoryEventStoreSuite struct {
-	store *MemoryEventStore
-	bus   *MockEventBus
-}
-
-func (s *MemoryEventStoreSuite) SetUpTest(c *C) {
-	s.bus = &MockEventBus{
-		events: make([]Event, 0),
-	}
-	s.store = NewMemoryEventStore(s.bus)
-}
-
-func (s *MemoryEventStoreSuite) Test_NewMemoryEventStore(c *C) {
-	bus := &MockEventBus{
-		events: make([]Event, 0),
-	}
-	store := NewMemoryEventStore(bus)
-	c.Assert(store, NotNil)
-}
-
-func (s *MemoryEventStoreSuite) Test_NoEvents(c *C) {
-	err := s.store.Save([]Event{})
-	c.Assert(err, Equals, ErrNoEventsToAppend)
-}
-
-func (s *MemoryEventStoreSuite) Test_OneEvent(c *C) {
-	event1 := &TestEvent{NewUUID(), "event1"}
-	err := s.store.Save([]Event{event1})
-	c.Assert(err, IsNil)
-	events, err := s.store.Load(event1.TestID)
-	c.Assert(err, IsNil)
-	c.Assert(events, HasLen, 1)
-	c.Assert(events[0], DeepEquals, event1)
-	c.Assert(s.bus.events, DeepEquals, events)
-}
-
-func (s *MemoryEventStoreSuite) Test_TwoEvents(c *C) {
-	event1 := &TestEvent{NewUUID(), "event1"}
-	event2 := &TestEvent{event1.TestID, "event2"}
-	err := s.store.Save([]Event{event1, event2})
-	c.Assert(err, IsNil)
-	events, err := s.store.Load(event1.TestID)
-	c.Assert(err, IsNil)
-	c.Assert(events, HasLen, 2)
-	c.Assert(events[0], DeepEquals, event1)
-	c.Assert(events[1], DeepEquals, event2)
-	c.Assert(s.bus.events, DeepEquals, events)
-}
-
-func (s *MemoryEventStoreSuite) Test_DifferentAggregates(c *C) {
-	event1 := &TestEvent{NewUUID(), "event1"}
-	event2 := &TestEvent{NewUUID(), "event2"}
-	err := s.store.Save([]Event{event1, event2})
-	c.Assert(err, IsNil)
-	events, err := s.store.Load(event1.TestID)
-	c.Assert(err, IsNil)
-	c.Assert(events, HasLen, 1)
-	c.Assert(events[0], DeepEquals, event1)
-	events, err = s.store.Load(event2.TestID)
-	c.Assert(err, IsNil)
-	c.Assert(events, HasLen, 1)
-	c.Assert(events[0], DeepEquals, event2)
-	c.Assert(s.bus.events, DeepEquals, []Event{event1, event2})
-}
-
-func (s *MemoryEventStoreSuite) Test_LoadNoEvents(c *C) {
-	events, err := s.store.Load(NewUUID())
-	c.Assert(err, ErrorMatches, "could not find events")
-	c.Assert(events, DeepEquals, []Event(nil))
-}
 
 type TraceEventStoreSuite struct {
 	baseStore *MemoryEventStore
